+++ src/Get/Anime.php
@@ -8,6 +8,7 @@ use Jikan\Lib\Parser\AnimeEpisodeParse;
 use Jikan\Lib\Parser\AnimeNewsParse;
 use Jikan\Lib\Parser\AnimeVideoParse;
 use Jikan\Lib\Parser\AnimeStatsParse;
+use Jikan\Lib\Parser\AnimePicturesParse;
 
 
 class Anime extends Get
@@ -15,7 +16,7 @@ class Anime extends Get
 
     public $canonical_path;
 
-    private $validExtends = [CHARACTERS_STAFF, EPISODES, NEWS, VIDEOS, STATS];
+    private $validExtends = [CHARACTERS_STAFF, EPISODES, NEWS, VIDEOS, STATS, PICTURES];
 
 	public function __construct($id = null, $extend = null) {
 
@@ -112,6 +113,15 @@ class Anime extends Get
         $this->parser = new AnimeStatsParse;
 
         $this->parser->setPath($this->canonical_path.'/stats');
+        $this->parser->loadFile();
+
+        $this->response = array_merge($this->response, $this->parser->parse());
+    }
+
+    private function pictures() {
+        $this->parser = new AnimePicturesParse;
+
+        $this->parser->setPath($this->canonical_path.'/pics');
         $this->parser->loadFile();
 
         $this->response = array_merge($this->response, $this->parser->parse());