+++ src/Model/AnimePictures.php
@@ -6,6 +6,6 @@ namespace Jikan\Model;
 class AnimePictures extends Model
 {
 
-	public $pictures = [];
+	public $image = [];
 
 }