+++ src/Model/CharacterPictures.php
@@ -6,6 +6,6 @@ namespace Jikan\Model;
 class CharacterPictures extends Model
 {
 
-	public $pictures = [];
+	public $image = [];
 
 }