+++ src/Model/MangaPictures.php
@@ -6,6 +6,6 @@ namespace Jikan\Model;
 class MangaPictures extends Model
 {
 
-	public $pictures = [];
+	public $image = [];
 
 }