+++ src/Model/PersonPictures.php
@@ -6,6 +6,6 @@ namespace Jikan\Model;
 class PersonPictures extends Model
 {
 
-	public $pictures = [];
+	public $image = [];
 
 }