+++ changelog.md
@@ -1,9 +1,11 @@
 ## Changelog
 
-### 1.6.3 stable - January 27, 18
+### 1.7.0 stable - January 27, 18
 - **[Anime|Manga]**
-	- Fix for [#98](/../../issues/98)
+	- Fix for [#98](/../../issues/98) - 1.6.3
 		- Return null for ISO8601 of aired_string that only has year
+- **[Anime|Manga|Character|Person]**
+	- Add gallery page parsing
 
 ### 1.6.2 stable - January 26, 18
 - **[Anime|Manga]**